+++ asapdiscovery-cli/asapdiscovery/cli/cli_args.py
@@ -1,7 +1,7 @@
 import click
 from asapdiscovery.data.dask_utils import DaskType
 from asapdiscovery.data.postera.manifold_data_validation import TargetTags
-from asapdiscovery.ml.models.ml_models import ASAPMLModelRegistry
+from asapdiscovery.ml.models import ASAPMLModelRegistry
 from asapdiscovery.modeling.protein_prep_v2 import CacheType
