+++ asapdiscovery-docking/asapdiscovery/docking/scorer_v2.py
@@ -16,7 +16,7 @@ from asapdiscovery.docking.docking_data_validation import (
 )
 from asapdiscovery.docking.docking_v2 import DockingResult
 from asapdiscovery.ml.inference import InferenceBase, get_inference_cls_from_model_type
-from asapdiscovery.ml.models.ml_models import MLModelType
+from asapdiscovery.ml.models import MLModelType
 from pydantic import BaseModel, Field
