+++ asapdiscovery-docking/asapdiscovery/docking/workflows/large_scale_docking.py
@@ -28,7 +28,7 @@ from asapdiscovery.docking.docking_data_validation import (
 )
 from asapdiscovery.docking.openeye import POSITDocker
 from asapdiscovery.docking.scorer_v2 import ChemGauss4Scorer, MetaScorer, MLModelScorer
-from asapdiscovery.ml.models.ml_models import ASAPMLModelRegistry
+from asapdiscovery.ml.models import ASAPMLModelRegistry
 from asapdiscovery.modeling.protein_prep_v2 import CacheType, ProteinPrepper
 from distributed import Client
 from pydantic import BaseModel, Field, PositiveInt, root_validator, validator
