--- asapdiscovery-ml/asapdiscovery/ml/loss/__init__.py
@@ -1,4 +1,0 @@
-from .GaussianNLLLoss import GaussianNLLLoss
-from .MSELoss import MSELoss
-
-__all__ = ["MSELoss", "GaussianNLLLoss"]
