+++ asapdiscovery-ml/asapdiscovery/ml/utils.py
@@ -1633,7 +1633,7 @@ def train(
     print("Using optimizer", optimizer, flush=True)
 
     if loss_fn is None:
-        from asapdiscovery.ml.loss.MSELoss import MSELoss
+        from asapdiscovery.ml.loss import MSELoss
 
         loss_fn = MSELoss()
     print("Using loss function", loss_fn, flush=True)
